Use 32 bit types for texture dimensions

+++ include/visionaray/texture/detail/prefilter.h
@@ -6,7 +6,6 @@
 #ifndef VSNRAY_TEXTURE_DETAIL_PREFILTER_H
 #define VSNRAY_TEXTURE_DETAIL_PREFILTER_H 1
 
-#include <cstddef>
 #include <stdexcept>
 
 #include <visionaray/math/detail/math.h>
@@ -26,17 +25,17 @@ namespace detail
 
 static float const Pole = sqrt(3.0f) - 2.0f;
 
-inline float init_causal_coeff(short* c, size_t len, size_t stride)
+inline float init_causal_coeff(short* c, unsigned len, unsigned stride)
 {
 
     typedef float float_type;
 //  typedef short voxel_type;
 
-    size_t const Horizon = min<size_t>(12, len);
+    unsigned const Horizon = min<unsigned>(12, len);
 
     float_type zk(Pole);
     float_type sum = float_type(*c) ;
-    for (size_t k = 0; k < Horizon; ++k)
+    for (unsigned k = 0; k < Horizon; ++k)
     {
         sum += zk * float_type(*c);
         zk *= float_type(Pole);
@@ -53,12 +52,12 @@ inline float init_anticausal_coeff(short* c)
 }
 
 template <typename T>
-inline void convert_to_bspline_coeffs(T*, size_t, size_t)
+inline void convert_to_bspline_coeffs(T*, unsigned, unsigned)
 {
     throw std::runtime_error("not implemented yet");
 }
 
-inline void convert_to_bspline_coeffs(short* c, size_t len, size_t stride)
+inline void convert_to_bspline_coeffs(short* c, unsigned len, unsigned stride)
 {
 
     typedef float float_type;
@@ -70,7 +69,7 @@ inline void convert_to_bspline_coeffs(short* c, size_t len, size_t stride)
 
     *c = voxel_type( Lambda * init_causal_coeff(c, len, stride) );
 
-    for (size_t k = 1; k < len; ++k)
+    for (unsigned k = 1; k < len; ++k)
     {
         c += stride;
         *c = voxel_type( Lambda * float_type(*c) + Pole * *(c - stride) );
@@ -106,14 +105,14 @@ inline void convert_for_bspline_interpol(texture_ref<T, 2> const* tex)
     using namespace detail;
 
     // row-wise
-    for (size_t row = 0; row < tex->height(); ++row)
+    for (unsigned row = 0; row < tex->height(); ++row)
     {
         T* ptr = &(*tex)[row * tex->width()]; // TODO: use operator()
         convert_to_bspline_coeffs(ptr, tex->width(), 1);
     }
 
     // column-wise
-    for (size_t col = 0; col < tex->width(); ++col)
+    for (unsigned col = 0; col < tex->width(); ++col)
     {
         T* ptr = &(*tex)[col]; // TODO: use operator()
         convert_to_bspline_coeffs(ptr, tex->height(), tex->width());
@@ -127,27 +126,27 @@ inline void convert_for_bspline_interpol(texture_ref<T, 3> const* tex)
 
     short* tmp = tex->prefiltered_data;
 
-    for (size_t z = 0; z < tex->depth(); ++z)
+    for (unsigned z = 0; z < tex->depth(); ++z)
     {
-        for (size_t y = 0; y < tex->height(); ++y)
+        for (unsigned y = 0; y < tex->height(); ++y)
         {
             short* ptr = &tmp[z * tex->width() * tex->height() + y * tex->width()];
             convert_to_bspline_coeffs(ptr, tex->width(), 1);
         }
     }
 
-    for (size_t x = 0; x < tex->width(); ++x)
+    for (unsigned x = 0; x < tex->width(); ++x)
     {
-        for (size_t z = 0; z < tex->depth(); ++z)
+        for (unsigned z = 0; z < tex->depth(); ++z)
         {
             short* ptr = &tmp[z * tex->width() * tex->height() + x];
             convert_to_bspline_coeffs(ptr, tex->height(), tex->width());
         }
     }
 
-    for (size_t y = 0; y < tex->height(); ++y)
+    for (unsigned y = 0; y < tex->height(); ++y)
     {
-        for (size_t x = 0; x < tex->width(); ++x)
+        for (unsigned x = 0; x < tex->width(); ++x)
         {
             short* ptr = &tmp[y * tex->width() + x];
             convert_to_bspline_coeffs(ptr, tex->depth(), tex->width() * tex->height());
